Fix failing test related to escaping include_blank in select_tag
Rails 3.0.x doesn't have the :prompt option in select_tag, it was
introduced in c5d54be746473ce5610e5e6de51f6b9d2495c935 that is only
available from 3.1.x on.

The test and related fix were introduced in
c9795871baacbe85ed4f19c2b830006fc6faeffd for Rails 3.0.17, as a fix for
a security vulnerability. The code is completely fine but the test was
using the invalid :prompt option for this version, probably because it
was cherry-picked from other branch which has the option.