Don't discard cache during asset precompile
A change was introduced in rails in 7f1a666d causing the whole application cache
to be cleared everytime a precompile is run, but it is not neccesary and just
slows down precompiling.

Secondary consequences are the clearing of the whole cache, which if using
the default file cache could cause an application level performance hit.

This is already [fixed in sprockets-rails][sprockets-rails] for rails 4, but is
a worthwhile point release patch for interim gains considering how much of a
pain point this is for so many.

  [sprockets-rails]: https://github.com/rails/sprockets-rails/blob/22270f83fb63537cef8a638835363360f2fbaab4/lib/sprockets/rails/task.rb