Revert "Switched to newer rdoc and gem package tasks (and their requires)."
This is a manual revert of commit 79aa54d, since the commit itself touches
in addition some version numbers.

API generation before Rails 3 uses the Jamis template, which requires an
old version of RDoc. To generate the API you need Rake 0.8.x or 0.9.x,
and the RDoc distributed with 1.8.7 (version 1.0.1).