Remove config.autoload_paths from config/application.rb
Most application developers shouldn't be adding paths to
`config.autoload_paths` as it won't be eagerly loaded so
remove it from the generated config/application.rb file.

If a developer does want to autoload from paths outside of the app
directory then they should add the path to `config.eager_load_paths`
as these paths are also autoloaded as well.