after_remove callbacks only run if record was removed.
If a record that does not belong to a collection is
passed in to a collection to be deleted, the after_remove
callback is run even though the record was not actually
removed. This can cause invalid behaviour, such as decrementing
a cache counter.

The new behaviour still has before_remove called, but after_remove
called only if the record was actually removed from the collection.