Do not redirect cache logger to /dev/null in test
For some reason, redirecting cache's logger to '/dev/null' resulting in
a test failures and LoadError. I think it's because of Thread issue.
Instead of trying to make every logger threadsafe for Rails 3.2, I think
it's better to just don't set the logger for now.

(Note: resetting the logger back to original value in the teardown block
didn't fix the problem.)