if force_ssl filters the redirected url loses the format
It happens because the format is not added to redirect_options in `force_ssl_redirect`
This PR add the format to redirected url if the format exists in Mime::Type
and if the `format` match the request.path and request.format.`format`?
responds true. In this way it could be works fine also with the `default` formats
in route

fixes #9061