Refactor index algorithm lookup so that it only builds the available options once
This way the available options are only built for actually fetching the
algorithm in case the option is given.

The options are going to be necessary a second time only in case the option
is given but does not exist, which is supposed to be due to a typo or
something like that, so no problem.