Allow serving of static files if relative_url_root is set
Currently static files aren't served if
Rails.application.config.relative_url_root is set to some path. This is
because the path in request is translated directly to a directory on file
system. With this patch, the relative_url_root path is first stripped
from the request so that static files are now found and served.