Raise helpful execptions on unexpected values.
When the attribute does exist, but we don't know what class to use for a
multi-parameter assignment, we now raise a helpful exception.

This should help people TDD their way to a correct implementation of
multi-parameter attributes, or make it more clear what's going on when there's a
real error.