Use Kernel#quietly when checking for pending migrations
The default logger Rails provides as ActiveRecord::Base.logger has a
method called `#silence` that takes no arguments. But ActiveSupport's
core extension requires a stream to be passed in, so it would break when
using a regular ::Logger instance.

Switched to using `Kernel#quietly`, which silences both STDOUT and
STDERR.