Fix #9127. Extract PostgreSQL array column defaults more reliable.
Cover all possible notations for array defaults when extracting
default value from column attributes.

As described in #9127 defaults of PostgreSQL columns of array types
are not extracted when the ARRAY constructor expression was used since
only the curly braces notation is expected.
This can lead to new instances of activerecord models not prefilling
the relating attribute with the default value as they should.