Don't touch test db when Rails.env == development
Currently, `rake db:create`, `rake db:drop`, and derived tasks (such as `rake
db:reset`) create and drop both the development and test databases when
`'development' == Rails.env`. Such behaviour vioates the Principle of Least
Surprise. It's extremely confusing to execute this:

    $ bundle exec rake db:reset

or this:

    $ RAILS_ENV=development bundle exec rake db:reset

...and discover that your test database is being recreated. This behaviour can
be functionally problematic. For example, assume a Rails project had the
following in their `database.yml` config:

    development:
      adapter: sqlite3
      database: db/development.sqlite3
      pool: 5
      timeout: 5000

    test:
      adapter: postgresql
      encoding: unicode
      database: myapp_test
      pool: 5
      username: username
      password: hackme

Recreating the test database is problematic for a developer who does not have
postgresql installed and configured. Rake tasks which attempt to recreate the
development database will fail because of the test database. That's silly.

This commit fixes this behaviour. The commit is very minimal: though it fixes
the behaviour described, the changes are not described as docstrings, and
existing method names are confusing.