name change active_record => base_klass
In Reflection instance active_record holds the information about
the base class.

The name `active_record` throws me off a little bit because the
name suggests that it holds an instance of the model class. While
it holds a model class itself.

This PR changes the name active_record => base_klass.