Store request.uuid in the current thread
This change to ActionDispatch::RequestId stores the
request's id in a slot on the current thread.

This lets included gems (or code that doesn't have
explicit access to the Rack env or ActionDispatch::Request)
forward the request's id to downstream services.

That allows for correlation of log entries across
services in a distributed system, which can be very
helpful in diagnosing issues.