Move the javascript include tag after the yield on layout/application.html.erb
Well, we all talk about performance, best practices and improvements. Moving
the include tag to the bottom will decrease the page load of your app just a bit.
It's just a way to start.

we always do this as a common practice to better performance, so why not do this
super simple change on the very first layout, it doesn't hurt anybody.