Removing not useful test related to reflect_on_all_associations
The comment in the test pretty much summarizes the issue.
`FIXME these assertions bust a lot`

Adding any type of association in class `Firm` will break this test.

I removed some deprecated stuff and this test failed.

I do not think this test provides any useful value. First of all who
counted last that 39 is the right number of associations.

Secondly there are a large number of tests which depend on reflection
returning right information about associations. Those tests will start
failing if there is a bug in the code.