Ensure args to Kernal.system are strings
The ActiveRecord rake tasks for dumping and loading the database
structure when using MySQL rely on calls to Kernel.system in order to
shell out to mysqldump and mysql. Kernel.system expects its arguments to
be Strings.

The Mysql2 adapter expects its port argument to be a Fixnum, so it
generally is given as such in database.yml. The values read from
database.yml should be cast to strings before being passed to
Kernel.system to keep everyone happy.