Wrap Minitest's assert_raise so that it assert on the message.
There are 14 assertions in the test suite that assert that a particular exception is raised and its message. Unfortunately Minitest don't test the message part. So all those tests are red herrings. This commit fixes it by wrapping the assert_raise method and make it compare the message after comparing the exception.