Add "extname" option to javascript_include_tag
ActionView::Helpers.asset_path is where the logic for
javascript_include_tag resides.  It takes an extname option for
specifying the extension or false to not append it.  This exposes that
option to javascript_include_tag.

Without the option files that didn't end with ".js" would get the
extension appended to them.  This broke JST templates and other file
types that should be interpreted as JavaScript but who's file extension
isn't ".js"