Fix ActiveSupport::Testing::Isolation on jruby
Hopefully the first of many related to #11700.

With these changes the railties tests wont 100% pass but at least they
will run and show the errors.

The first problem was the @method_name variable was never set.
This was causing ENV["ISOLATION_TEST"] to always be nil which
would cause the test output to never be written to the tempfile.

The second problem was that an invalid -t option was being passed
to minitest and minitest was erroring out.

The third problem was the run method needs to always return an
instance of the test class it is trying to isolate. We were returning
a ProxyTestResult instead.