Change generated README file to use Markdown
Rails guides are now markdown as well as the main README for the rails project. The generated README when running `rails new` should be the same format.

Prior to this PR the generated README for both the main app `rails new` and a plugin `rails plugin new` generated a `README.rdoc`. With this PR both will instead generate a `README.md`.