Conform to RFC2616-sec14 for */* media-range, #8987
When requesting:

    Accept: application/json, text/javascript, */*; q=0.01

That q=0.01 marks */* with the least preference. We should not
short-circuit the Mime::Mimes sorting, which is handling the q-values
correctly.

This fixes the default request by jQuery.ajax(url, {dataType:'json'}),
which makes a request with the above Accept header, but Rails would
respond with HTML instead of JSON.