Revert "add active model basics wip to guides/documents.yaml"
Reason: WIP guides are not in the index because they are partial
work pushed to the repo. Either there is someone working on them
or else the work was interrupted.

Guides are added to the index page when they are finished, the
author considers the draft to be complete, someone reviews them,
and overall they are considered to be good for users to read.

This reverts commit eefc03a8069bddfe31751c4bd4fb1804e7f33c2c.