Handle errors in SQL queries with UTF8-incompatible encodings.
When an application issues a SQL query whose encoding isn't compatbile
with UTF-8, and the database returns an error, the logging code raises
an encoding exception.

After this commit, the error and SQL query will be converted to the
application's internal encoding (most likely UTF-8) in response to
encoding exceptions.

Fixes #5738.