allow for arrays to be defined as symbol
`change_colum :table, :column, :bigint, array: true` didn't work because
the array option tried to string concat: `:bigint << '[]'`

I don't know if this is the right place to convert the possible symbol
to a string. If it's wrong please let me know a better place and I will
change this.