fix retrieving duplicated columns with mysql
When a SQL query calls twice the same column with the same name without
a statement, mysql will loop through all the columns to define their
type.

Since there are two columns with the same name, the second time we
define it in the types array, we're just gonna override it without
adding a new entry.

Since the fields parser then expects one more column, the following ones
are offset from one another.