Merge pull request #13455 from jetthoughts/13445_fix_touch_destroyed_record
On destroying do not touch destroyed belongs to association.
Conflicts:
	activerecord/CHANGELOG.md
	activerecord/lib/active_record/associations/builder/belongs_to.rb

The code that raises an error when touching destroyed records is not on `4-0-stable`.
It's only on `master`: e52ff809606323f7370cfaf6e63b083f9b108d6c

I modified the CHANGELOG entry to explain the bug rather the exception raising behavior.
I also had to modify the test-case to use the same instance. Otherwise `destroy` did
not result on a `destroyed?` association record.

/cc @pftg