Render select tag before hidden input
Prior to this, if you render a select tag with the form helper you end up
with a idden field followed by the select tag. This is fine in most cases,
but on older versions of IE, when you are wrapping said select tag with a
label tag, the label will get associated with the hidden field rather
than the select so a click on the label will have no effect as opposed
to focusing on the select itself.