We want to allow aria-* to pass like data-* in tag options
Exactly like `data-*` attributes on HTML elements it is now common to
pass `aria-*` style attributes to tags (bootstrap for example). The
spec is defined here
http://rawgithub.com/w3c/aria-in-html/master/index.html

In this PR I’ve cleaned up some of the previous tag option formatting
to be a bit faster, smaller to reason, and opens up to future additions
of similar hash-like options. If for instance we wanted to add the
`rails-*` hash-like we’d just change `HASHLIKE_ATTRIBUTE_KEYS` to
`[‘data’, ‘aria’, ‘rails’]`